Move spec files and use Jasmine for browser and renderer specs
