+++ spec/helpers/createComponent.js
