+++ spec/insight-spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Insight = require('../../src/renderer/utils/insight.js');
+var Insight = require('../src/renderer/utils/insight.js');
 
 describe('Insight', function () {
